chore: move function 'getAdvancedSearchIdFromLocation' to AdvancedSearchModule
